Enhanced SEO
Created new metadatas and changed the layout
