+++ src/app/[locale]/layout.js
@@ -1,10 +1,9 @@
-import ClientLayout from "./ClientLayout";
-import { AuthProvider } from "../contexts/AuthContext";
-import { NextIntlClientProvider } from 'next-intl';
 import { getMessages } from 'next-intl/server';
 import { getSysVar } from "../../utils/FirestoreData";
 
-import "@/app/globals.css"
+export const metadata = {
+  title: process.env.NEXT_PUBLIC_NAME
+};
 
 async function RootLayout({
   children,
@@ -15,7 +14,9 @@ async function RootLayout({
   // Get global variables
   const sysvar = await getSysVar();
   return (
-    <html lang={loc.locale}>
+    <>
+    {children}
+    {/*<html lang={loc.locale}>
       <head>
         <link rel="icon" href="/logo.svg" />
         <meta name="theme-color" content="#000000" />
@@ -34,15 +35,10 @@ async function RootLayout({
         />
       </head>
       <body>
-        <NextIntlClientProvider locale={loc.locale} messages={messages}>
-            <AuthProvider sysVar={sysvar.system}>
-              <ClientLayout>
-                {children}
-              </ClientLayout>
-            </AuthProvider>
-        </NextIntlClientProvider>
+        {children}
       </body>
-    </html>
+    </html>*/}
+    </>
   );
 }
