+++ src/app/[locale]/login/page.js
@@ -1,5 +1,5 @@
 "use client"
-import React, { useEffect, useState, useTransition } from "react";
+import React, { useState } from "react";
 import LoginForm from "./components/form";
 import { useAuth } from "@/app/contexts/AuthContext";
 import { FirebaseLogin } from "@/app/hooks/FirebaseLogin";
